requirements.txt
Django>=5.1
pillow>=11.1.0
djangorestframework>=3.15.2
djangorestframework_simplejwt>=5.5.0